requirements_dsu.txt
# packages
django==2.2.1
urllib3==1.24.3
boto3==1.14.13
boto==2.49.0
django-storages==1.7.1
raven==6.10.0
enum34==1.1.6
requests-kerberos==0.12.0
mock==3.0.5

# only our third party here
ib_common==2.0.1
django-swagger-utils==2.2.0

# logentries
r7insight_python==0.9.2
Logentries==0.17

# cors-headers
django-cors-headers==3.0.1

# ib_sentry wrapper
ib_sentry_wrapper==1.0.2

schema==0.7.2

coverage==4.5.3
pytest==4.6.3
pytest-django==3.5.0
pytest-cov==2.7.1
freezegun==0.3.12
pytest-sugar==0.9.4
factory-boy==2.12.0
pytest-mock==3.2.0
